-> Refactoring of OffHeap version

+++ core/src/main/java/org/mwdb/chunk/KStack.java
@@ -4,7 +4,7 @@ public interface KStack {
 
     boolean enqueue(int index);
 
-    int dequeueTail();
+    long dequeueTail();
 
     boolean dequeue(int index);
